Fix: User 도메인 2개를 하나로 수정
