+++ src/main/java/com/hoodie/otti/model/profile/User.java
@@ -1,27 +1,37 @@
 package com.hoodie.otti.model.profile;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
 
 @Entity
-public class UserProfile {
+public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @NotEmpty(message = "닉네임은 필수 입력 사항입니다.")
+    @Size(min = 3, max = 50, message = "닉네임은 최소 3자에서 최대 50자여야 합니다.")
     private String username;
+
+    @NotEmpty(message = "이메일은 필수 입력 사항입니다.")
+    @Email(message = "올바른 이메일 형식이어야 합니다.")
+    private String userEmail;   // 사용자 이메일
+
     private String profilePhotoUrl;
 
-    public UserProfile() {
+    public User() {
     }
 
-    public UserProfile(String username, String profilePhotoUrl) {
+    public User(String username, String profilePhotoUrl, String userEmail) {
         this.username = username;
         this.profilePhotoUrl = profilePhotoUrl;
+        this.userEmail = userEmail;
     }
 
     public Long getId() {
