+++ src/main/java/com/hoodie/otti/repository/profile/UserRepository.java
@@ -1,0 +1,14 @@
+package com.hoodie.otti.repository.profile;
+
+import com.hoodie.otti.model.profile.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findById(Long UserId);
+
+    Optional<User> findByUserEmail(String userEmail);
+}
